% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_life_expectancy.r
\name{calculate_life_expectancy}
\alias{calculate_life_expectancy}
\title{calculate_life_expectancy}
\usage{
calculate_life_expectancy(age, a, b, vector_of_coefficients,
  log_hazard_ratio, data_for_weights, working_directory)
}
\arguments{
\item{age}{age}

\item{a}{- intercept of Gompertz baseline hazards}

\item{b}{- slope of Gompertz baseline hazards}

\item{vector_of_coefficients}{- vector of coefficients from cox-model}

\item{log_hazard_ratio}{- log hazard rations from cox-model}

\item{data_for_weights}{- weights, i.e proportions for each generated combination}

\item{working_directory}{- the working directory for any data that is produced from the functions to be saved in a folder}
}
\value{
data frame with life expectancies for given data frame of clients
calculate_life_expectancy()
}
\description{
This function calculates life expectancy table based on inputs
}
\keyword{life_expectancy}
