% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/function_for_table_and_survival_of_combination.r
\name{function_for_table_of_combination}
\alias{function_for_table_of_combination}
\title{function_for_table_of_combination
This function produces table with different combinations of scenarious and life expectancies}
\usage{
function_for_table_of_combination(
  age,
  a,
  b,
  vector_of_coefficients,
  LogHazardRations,
  data_for_weights,
  working_directory
)
}
\arguments{
\item{age}{age}

\item{a}{- intercept of Gompertz baseline hazards}

\item{b}{- slope of Gompertz baseline hazards}

\item{vector_of_coefficients}{- vector of coefficients from cox-model}

\item{LogHazardRations}{- log hazard rations from cox-model}

\item{data_for_weights}{- weights, i.e proportions for each generated combination}

\item{working_directory}{- the working directory for any data that is produced from the functions to be saved in a folder}
}
\value{
data frame with life expectancies for given data frame of clients
function_for_table_of_combination()
}
\description{
function_for_table_of_combination
This function produces table with different combinations of scenarious and life expectancies
}
\keyword{life_expectancy}
