% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikGenNPH.r
\name{LikGenNPH}
\alias{LikGenNPH}
\title{LikGenNPH
This function calculates the neglikelihood for parametrical survival model.}
\usage{
LikGenNPH(b, D, nf, nk, ncl, dist)
}
\arguments{
\item{b}{shape parameters defining the hazard functions}

\item{D}{The data set as described in previous section;}

\item{nf}{The number of continuous and binary factors in the data set D corresponding to the covariates used in the Cox-regression for proportional hazard term;}

\item{nk}{The number of continuous and binary factors in the data set D corresponding to the covariates used in the Cox-regression for shape b;}

\item{ncl}{The number of clusters in the data set D corresponding to the covariate defining the shared frailty (is equal to 0 for the fixed-effect model);}

\item{dist}{distribution either Weibull or Gompertz}
}
\description{
LikGenNPH
This function calculates the neglikelihood for parametrical survival model.
}
\keyword{for}
\keyword{model}
\keyword{neglikelihood}
\keyword{parametrical}
\keyword{survival}
\keyword{the}
