% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylongevity_method2.r
\name{mylongevity_method2}
\alias{mylongevity_method2}
\title{my longevity method2 function
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
Then using the argument 'list_of_variables' this functions selects the columns of interest to be matched
with coefficients (i.e log-hazard ratios) estimated in analysis}
\usage{
mylongevity_method2(data, indexes_of_variables, list_of_variables, age, a,
  b, working_directory)
}
\arguments{
\item{data}{given data frame of clients,}

\item{indexes_of_variables}{indices for columns of interest}

\item{list_of_variables}{columns of interest to be matched to coefficients (i.e log-hazard ratios)}

\item{age}{age of clients}

\item{a}{intercept from Gompertz baseline hazards}

\item{b}{slope from Gompertz baseline hazards}

\item{working_directory}{the working directory for any data that is produced from the functions to be saved in a folder}
}
\value{
data frame with life expectancies for given data frame of clients
mylongevity_method2()
}
\description{
my longevity method2 function
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
Then using the argument 'list_of_variables' this functions selects the columns of interest to be matched
with coefficients (i.e log-hazard ratios) estimated in analysis
}
\keyword{life_expectancy}
