% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylongevity_method3.r
\name{mylongevity_method3}
\alias{mylongevity_method3}
\title{my longevity method3 function
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
Then using the argument 'list_of_variables' this functions selects the columns of interest to be matched
with coefficients (i.e log-hazard ratios) estimated in analysis}
\usage{
mylongevity_method3(data, indexes_of_variables, list_of_variables, age, a,
  b, T_start_indicator, T_stop_indicator, status_indicator,
  working_directory)
}
\arguments{
\item{data}{given data frame of clients}

\item{indexes_of_variables}{indices for columns of interest}

\item{list_of_variables}{columns of interest to be matched to coefficients (i.e log-hazard ratios)}

\item{age}{age of clients}

\item{T_start_indicator}{for right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval}

\item{T_stop_indicator}{ending time of the interval for interval censored or counting process data only. Intervals are assumed to be open on the left and closed on the right, (start, end]. For counting process data, event indicates whether an event occurred at the end of the interval}

\item{status_indicator}{the status indicator, normally 0=alive, 1=dead}

\item{working_directory}{the working directory for any data that is produced from the functions to be saved in a folder}
}
\value{
data frame with life expectancies for given data frame of clients
mylongevity_method3()
}
\description{
my longevity method3 function
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
Then using the argument 'list_of_variables' this functions selects the columns of interest to be matched
with coefficients (i.e log-hazard ratios) estimated in analysis
}
\keyword{life_expectancy}
