% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParNPHCox.r
\name{ParNPHCox}
\alias{ParNPHCox}
\title{ParNPHCox}
\usage{
ParNPHCox(formula.scale, formula.shape, cluster, dist, data)
}
\arguments{
\item{formula.scale}{The formula object defining the fields for time-to-failure (or timeof-start and time-to-failure) and for covariates influencing the proportional hazard term;}

\item{formula.shape}{The object defining the fields for covariates influencing the shape;}

\item{cluster}{The name of the covariate defining the random effect (is equal to NULL for the fixed-effect model);}

\item{dist}{distribution of the time-to-failure (’Weibull’ or ’Gompertz’);}

\item{data}{The data set of interest
This function calculates the estimates of unknown parameters, their standard errors, and other attributes.}
}
\description{
ParNPHCox
}
\keyword{The}
\keyword{and}
\keyword{attributes.}
\keyword{errors,}
\keyword{estimates}
\keyword{of}
\keyword{other}
\keyword{parameters,}
\keyword{standard}
\keyword{their}
\keyword{unknown}
