% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylongevity_double_cox_method1.r
\name{mylongevity_double_cox_method1}
\alias{mylongevity_double_cox_method1}
\title{mylongevity_double_cox_method1}
\usage{
mylongevity_double_cox_method1(
  data,
  indexes_of_variables,
  age_of_diagnosis,
  time_past_from_diagnosis
)
}
\arguments{
\item{data}{given data frame of clients}

\item{indexes_of_variables}{indices for columns of interest.}

\item{age_of_diagnosis}{age of diagnosis for type 2 diabetes mellitus}

\item{time_past_from_diagnosis}{time past from diagnosis in years from the date of type 2 diabetes mellitus}
}
\value{
data frame with life expectancies for given data frame of clients
res1<-mylongevity_double_cox_method1(data,indexes_of_variables, age_of_diagnosis,time_past_from_diagnosis)
}
\description{
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
and matches them to outcomes of hazard ratios from type 2 diabetes study
}
\details{
This method matches the attributes of inputted data with results of hazard ratios from type 2 diabetes study and produces a table with life expectancies
For mylongevity_double_cox_method1, user has to specify the data and columns using the argument indexes_of_variables in following order: atrial,birth_year,bmi_regrp,group,hf,hyperchol,hypert,miocarInfarct,pvd,sex,smokes,townsend
The definition of columns are:
   -	atrial: indicator for atrial fibrillation  (0 - absense of atrial fibrillation or 1 - presence of atrial fibrillation)
   -	birth_year  indicator for year of birth   (0 - born in 1930-1939, 1 - born in 1950-1960, 2 - born in  1940-1949)
   -	bmi_regrp : indicator for body mass index (0 - Healthy, 1 - Overweight, 2 - Obese)
   -	group : indicator of cases control group (0 - absense type 2 diabetes mellitus, 1 - presence type 2 diabetes mellitus)
   -	hf : indicator for heart failure (0 - absense of heart failure, 1 - presence of heart failure)
   -	hyperchol: indicator for hypercholesterolemia (0 - absense of hypercholesterolemia, 1 - presence of hypercholesterolemia)
   -	hypert: indicator for hypertension (0 - absense of Hypertension, 1 - Treated Hypertension 2 - Untreated Hypertension)
   -	miocarInfarct: indicator for myocardial infarction (0 - absense of myocardial infarction, 1 - presence of myocardial infarction)
   -	pvd: indicator for peripheral vascular disease (0 - absense of peripheral vascular disease, 1 - presence of peripheral vascular disease)
   -	sex: indicator for gender (0 - Female, 1 - Male)
   -	smokes: indicator of smoking status (0 non smoker, 1 - former smoker, 2 - current smoker)
   -	townsend - deprivation index (0 - (Townsend 3),  1 - (Townsend 1)  2 - (Townsend 2) , 3 - (Townsend 4), 4 - (Townsend 5))
}
\keyword{life_expectancy}
