% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylongevity_double_cox_method2.r
\name{mylongevity_double_cox_method2}
\alias{mylongevity_double_cox_method2}
\title{mylongevity_double_cox_method2}
\usage{
mylongevity_double_cox_method2(
  data,
  indexes_of_variables,
  age_of_diagnosis,
  time_past_from_diagnosis
)
}
\arguments{
\item{data}{given data frame of clients}

\item{indexes_of_variables}{indices for columns of interest}

\item{age_of_diagnosis}{age of start of intake for hormone replace therapy}

\item{time_past_from_diagnosis}{time past from start of hormone replace therapy in years}
}
\value{
data frame with life expectancies for given data frame of clients
}
\description{
This function loads a file as a data frame of clients.
Using the argument 'indexes_of_variables', this function selects the columns of factors
and matches them to outcomes of life expectancy calculator from hormone replace therapy study
}
\details{
This method matches the attributes of inputted data with results of hazard ratios from hormone replace therapy study and produces a table with life expectancies
For mylongevity_double_cox_method2, user has to specify the data and columns using the argument indexes_of_variables in following order: B_cohort,BMIsmoke,CHD,deprivation,DMIIsmoke,hrtcat,hypertension,opho

The definition of columns are:
   -	B_cohort : indicator for year of birth  (0 - born in 1921-1930 , 1 - born in 1931-1940,  2 - born in  1941-1950, 3 - born in 1951-1960)
   -	BMIsmoke : indicator for interaction for body mass index and smoking (0 - Healthy or Overweight and Non Smoker, 1 - healthy weight and current smoker, 2 - healthy weight and ex smoker, 3, Obese and current smoker, 4 - Obese and exsmoker, 5 - Obese and non smoker)
   -	CHD : indicator for heart failure (0 - absence of coronary heart disease , 1 - presence of coronary heart disease)
   - hrtcat : indicator for type of hormone replacement therapy (0 - non users of hormone replacement therapy, 1  - Combined Oestrogen and Progesteron , 2 - Oestrogen-only)
   - DMIIsmoke : interaction of diabetes and smoking variables (0 -  non diabetic and non smoker,1 - non diabetic  and current smoker ,2 - non diabetic and ex-smoker,3 - diabetic and current smoker,4 - diabetic  and ex smoker ,5 - diabetic and non-smoker)
   -	opho : indicator of cases control group (1 - both uterus and ovarian removed, 2 - ovaries removed but not uteras)
   -	hypertension: indicator for hypertension (0 - absence of Hypertension, 1 -  Hypertension)
   -	smokes: indicator of smoking status (0 non smoker, 1 - former smoker, 2 - current smoker)
   -	deprivation - deprivation index (0 - Low (Townsend 1 and Townsend 2), 1 - High (Townsend 4 and Townsend 5), 2 - Medium (Townsend 3))
}
\keyword{life_expectancy}
